+++ example/client.js
@@ -7,6 +7,8 @@ var disque = new Disqueue( {
 	//'auth' : {
 		//'password' : 'gideonairex'
 	//}
+	'retryMax'   : 5,
+	'retryDelay' : 2000
 } );
 
 disque.on( 'error', function ( error ) {
@@ -14,23 +16,22 @@ disque.on( 'error', function ( error ) {
 } );
 
 disque.on( 'connected', function () {
+	console.log( 'connected' );
+} );
 
-	var server = http.createServer( function ( request, reply ) {
+var server = http.createServer( function ( request, reply ) {
+	disque.addJob( {
+		'queue' : 'test',
+		'job'   : 'Hello world',
+		'delay' : 2,
+		'ttl'   : 3
+	}, function ( error, data ) {
 
-		disque.addJob( {
-			'queue' : 'test',
-			'job'   : 'Hello world',
-			'delay' : 2,
-			'ttl'   : 3
-		}, function ( error, data ) {
-
-			console.log( error );
-			reply.end( data );
-
-		} );
+		console.log( error );
+		reply.end( data );
 
 	} );
 
-	server.listen( 9991 );
-
 } );
+
+server.listen( 9991 );