+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+### 1.4.0
+* Add qscan command
+
+### 1.3.1
+* Update uuid to version 3.0.0
+
+### 1.3.0
+* Add configurable retry
+
+### 1.2.0
+* Add working command
+
 ### 1.1.0
 * Add decorators from disqueu raw
 * Complete tests for existing commands